+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## ****Unreleased****
 
+#### Added
+- Support floating point special values
+
 ## [1.9.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.8.0...1.9.0) (2023-08-03)
 
 #### Added